Refactor S2871 false positive matchers
